% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_batch_2_pg_grskey_grid.R
\name{add_batch_2_pg_grskey_grid}
\alias{add_batch_2_pg_grskey_grid}
\title{Update FAIB hectares database from input csv of datasets}
\usage{
add_batch_2_pg_grskey_grid(
  inCSV = "D:\\\\Projects\\\\provDataProject\\\\tools\\\\prov_data_resultant3.csv",
  connList = faibDataManagement::get_pg_conn_list(),
  oraConnList = faibDataManagement::get_ora_conn_list(),
  cropExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  gr_skey_tbl = "all_bc_gr_skey",
  wrkSchema = "whse",
  rasSchema = "raster",
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Lands_and_Islandsincluded.tif",
  dataSourceTblName = "data_sources",
  setwd = "D:/Projects/provDataProject",
  outTifpath = "D:\\\\Projects\\\\provDataProject",
  importrast2pg = FALSE
)
}
\arguments{
\item{inCSV}{coming soon}

\item{connList}{coming soon}

\item{cropExtent}{coming soon}

\item{gr_skey_tbl}{coming soon}

\item{wrkSchema}{coming soon}

\item{rasSchema}{coming soon}

\item{dataSourceTblName}{coming soon}

\item{setwd}{coming soon}

\item{templateRaster}{coming soon}
}
\value{
no return
}
\description{
Update FAIB hectares database from input csv of datasets
}
\examples{
coming soon
}
