% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_2_pg_grskey_grid.R
\name{add_data_2_pg_grskey_grid}
\alias{add_data_2_pg_grskey_grid}
\title{Update FAIB hectares database from input dataset}
\usage{
add_data_2_pg_grskey_grid(
  rslt_ind,
  srctype,
  srcpath,
  srclyr,
  suffix,
  nsTblm,
  query,
  flds2keep,
  connList = faibDataManagement::get_pg_conn_list(),
  oraConnList = faibDataManagement::get_ora_conn_list(),
  cropExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  gr_skey_tbl = "all_bc_res_gr_skey",
  wrkSchema = "whse",
  rasSchema = "raster",
  fdwSchema = "load",
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  dataSourceTblName = "data_sources",
  outTifpath = "D:\\\\Projects\\\\provDataProject",
  importrast2pg = FALSE
)
}
\arguments{
\item{rslt_ind}{Used in combination with gr_skey_tbl and suffix arguments. 1 = include (i.e. will add primary key to gr_skey tbl) 0 = not included (i.e. will not add primary key to gr_skey table)}

\item{srctype}{Format of data source, options: gdb, oracle, postgres, geopackage, raster, shp}

\item{srcpath}{Path to input data. Note use bcgw for whse oracle layers}

\item{srclyr}{Input layer name}

\item{suffix}{Use in combination with rslt_ind and gr_skey_tbl. Suffix to be added to the primary key field name in the gr_skey table}

\item{nsTblm}{Name of output non spatial table}

\item{query}{Where clause used to filter input dataset, Eg. fire_year = 2023 and BURN_SEVERITY_RATING in ('High','Medium')}

\item{flds2keep}{Fields to keep in non spatial table, Eg. fid,tsa_number,thlb_fact,tsr_report_year,abt_name}

\item{connList}{Keyring object of Postgres credentials}

\item{cropExtent}{Raster crop extent, list of c(ymin, ymax, xmin, xmax) in EPSG:3005}

\item{gr_skey_tbl}{Pre-existing table that contains gr_skey field}

\item{wrkSchema}{Schema to insert the new gr_skey and non spatial table}

\item{rasSchema}{If importrast2pg set to TRUE, the import schema for the raster}

\item{fdwSchema}{If srctype=oracle, the schema to load the foreign data wrapper table}

\item{grskeyTIF}{The file path to the gr_skey geotiff}

\item{maskTif}{The file path to the geotiff to be used as a mask}

\item{dataSourceTblName}{coming soon}

\item{outTifpath}{Directory where output tif if exported and where vector is temporally stored prior to import}

\item{importrast2pg}{If TRUE, raster is imported into database in rasSchema}
}
\value{
no return
}
\description{
Update FAIB hectares database from input dataset
}
\examples{
coming soon
}
