% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_2_pg_grskey_with_overlap.R
\name{add_data_2_pg_grskey_with_overlap}
\alias{add_data_2_pg_grskey_with_overlap}
\title{Import data with spatial overlaps into Postgres as gr_skey table}
\usage{
add_data_2_pg_grskey_with_overlap(
  inSrc,
  inLyr,
  groupFld,
  outPGTblName,
  schema = "whse",
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Lands_and_Islandsincluded.tif",
  connList = get_pg_conn_list(),
  terraExt = terra::ext(c(273287.5, 1870587.5, 367787.5, 1735787.5))
)
}
\arguments{
\item{grskeyTIF}{coming soon}

\item{maskTif}{coming soon}

\item{connList}{coming soon}

\item{rslt_ind}{coming soon}

\item{srctype}{coming soon}

\item{srcpath}{coming soon}

\item{srclyr}{coming soon}

\item{pk}{coming soon}

\item{suffix}{coming soon}

\item{nsTblm}{coming soon}

\item{query}{coming soon}

\item{flds2keep}{coming soon}

\item{cropExtent}{coming soon}

\item{gr_skey_tbl}{coming soon}

\item{wrkSchema}{coming soon}

\item{rasSchema}{coming soon}

\item{dataSourceTblName}{coming soon}

\item{setwd}{coming soon}
}
\value{
no return
}
\description{
Import data with spatial overlaps into Postgres as gr_skey table
}
\examples{
coming soon
}
