% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-fields-to-resultant.R
\name{add_fields_to_resultant}
\alias{add_fields_to_resultant}
\title{Update the resultant table with fields from previously imported datasets.}
\usage{
add_fields_to_resultant(
  resultant_table_output_name,
  join_gr_skey_table,
  fields_to_include,
  join_attribute_table = NULL,
  overwrite_fields = FALSE,
  resultant_table = "resultant",
  pg_conn_param = dadmtools::get_pg_conn_list(),
  key_resultant_tbl = "gr_skey",
  key_grskey_tbl = "gr_skey",
  key_join_tbl = "pgid",
  updated_field_names = NULL,
  include_prefix = FALSE,
  prefix = NULL,
  overwrite_resultant_table = FALSE
)
}
\arguments{
\item{resultant_table_output_name}{user selected name of the output resultant table. Note: Include schema in the name.}

\item{join_gr_skey_table}{the name of the table containing the key (e.g. gr_skey) to join to resultant key.}

\item{fields_to_include}{vector of fields to include from joining tables.}

\item{join_attribute_table}{the name of the attribute table containing the key (e.g. pgid) to join to gr_skey table. Attributes from this table will be in the final resultant table. If the table name is left blank, gr_skey table attributes will be used in the final resultant table.}

\item{overwrite_fields}{DEFAULT:FALSE. TRUE = New Fields will overwrite any fields that already exists in the current Resultant table. FALSE = Existing Resultant table fields will not be overwritten.}

\item{resultant_table}{name of the existing resultant table.}

\item{pg_conn_param}{named list of postgres connection parameters (see get_pg_conn_list() function for more details)}

\item{key_resultant_tbl}{DEFAULT:'gr_skey'. join key in resultant table (e.g. gr_skey)}

\item{key_grskey_tbl}{DEFAULT:'gr_skey'. join key in gr_key table (e.g. gr_skey)}

\item{key_join_tbl}{DEFAULT:'pgid'. join key in attribute table (e.g. pgid)}

\item{updated_field_names}{OPTIONAL vector of updated fields names to include int final resultant table.}

\item{include_prefix}{DEFAULT:FALSE. TRUE = a prefix will be added to new field names. FALSE = no prefix will be added to field names.}

\item{prefix}{the prefix to be added to the beginning of new field names}

\item{overwrite_resultant_table}{DEFAULT:FALSE. TRUE = any table with the resultant_table_output_name will be overwritten. FALSE = Resultant tbale will not be overwritten.}
}
\value{
nothing is returned
}
\description{
Update the resultant table with fields from previously imported datasets.
}
\examples{
coming soon
}
