% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-pk-to-gr-skey-tbl.R
\name{add_pk_to_gr_skey_tbl}
\alias{add_pk_to_gr_skey_tbl}
\title{Update the gr_skey table with primary keys from input table}
\usage{
add_pk_to_gr_skey_tbl(
  dst_tbl,
  dst_schema,
  pk,
  suffix,
  gr_skey_tbl,
  pg_conn_param,
  join_field = "gr_skey"
)
}
\arguments{
\item{dst_tbl}{input pg table to be incorporated into gr_skey table}

\item{dst_schema}{input pg table schema to be incorporated into gr_skey table}

\item{pk}{column to incorporated into gr_skey table}

\item{suffix}{suffix to be appended to the pk for the column name in the gr_skey table}

\item{gr_skey_tbl}{current gr_skey table}

\item{pg_conn_param}{named list of postgres connection parameters (see get_pg_conn_list() function for more details)}

\item{join_field}{defaults to 'gr_skey')}
}
\value{
nothing is returned
}
\description{
Update the gr_skey table with primary keys from input table
}
\examples{
coming soon
}
