% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-import-to-pg-gr-skey.R
\name{batch_import_to_pg_gr_skey}
\alias{batch_import_to_pg_gr_skey}
\title{Update FAIB hectares database from input csv of datasets}
\usage{
batch_import_to_pg_gr_skey(
  in_csv = "config_parameters.csv",
  pg_conn_param = faib_dadm_tools::get_pg_conn_list(),
  ora_conn_param = faib_dadm_tools::get_ora_conn_list(),
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  gr_skey_tbl = "whse.all_bc_gr_skey",
  dst_schema = "whse",
  raster_schema = "raster",
  template_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  mask_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  data_src_tbl = "whse.data_sources",
  out_tif_path,
  import_rast_to_pg = FALSE
)
}
\arguments{
\item{in_csv}{File path to data sources csv, see config_parameters.csv for example}

\item{pg_conn_param}{Keyring object of Postgres credentials}

\item{ora_conn_param}{Keyring object of Oracle credentials}

\item{crop_extent}{list of c(ymin, ymax, xmin, xmax) in EPSG:3005}

\item{gr_skey_tbl}{Schema and table name of the pre-existing foreign key lookup gr_skey table. Argument to be used with suffix and rslt_ind within in_csv.}

\item{raster_schema}{If importrast2pg=TRUE, schema of imported raster}

\item{template_tif}{The file path to the gr_skey geotiff to be used as a template raster}

\item{mask_tif}{The file path to the geotiff to be used as a mask}

\item{data_src_tbl}{Schema and table name of the metadata table in postgres that updates with any newly imported layer}

\item{out_tif_path}{Directory where output tif if exported and where vector is temporally stored prior to import}

\item{import_rast_to_pg}{If TRUE, raster is imported into database in rasSchema}

\item{working_schema}{Schema where gr_skey, non spatial table will be imported as well as the schema of gr_skey_tbl}
}
\value{
no return
}
\description{
Update FAIB hectares database from input csv of datasets
}
\examples{
coming soon
}
