% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-pg-to-seles-tif.R
\name{batch_pg_to_seles_tif}
\alias{batch_pg_to_seles_tif}
\title{Batch imports Postgres table fields (data types: integer | double | text) to TIFF.
Converts to an integer TIF. Double data types (decimals) are multiplied by the parameter d_mulitiplier, rounded and the result converted to an integer.
Text is converted to a text:sequence pair and saved to a CAT file, and the sequence is converted to the TIF.}
\usage{
batch_pg_to_seles_tif(
  in_csv = "batch_pg_to_seles_tif.csv",
  pg_conn_param = get_pg_conn_list(),
  template_tif,
  out_tif_path,
  out_cat_path,
  gdb,
  bnd
)
}
\arguments{
\item{in_csv}{File path to input configuration csv, defaults to "batch_pg_to_seles_tif.csv"}

\item{pg_conn_param}{Keyring object of Postgres credentials, defaults to dadmtools::get_pg_conn_list()}

\item{template_tif}{(required): string - file path to template tif (e.g. skey.tif)}

\item{out_tif_path}{(required): string - path to output rasters directory}

\item{out_cat_path}{(required): string - path to output cat files directory. Only used for text field types. Default NULL.}

\item{gdb}{(required): string - path to GDB that contains the unit boundary feature class (e.g. bnd) - used to create extent}

\item{bnd}{(required): string - boundary feature class name}
}
\description{
Batch imports Postgres table fields (data types: integer | double | text) to TIFF.
Converts to an integer TIF. Double data types (decimals) are multiplied by the parameter d_mulitiplier, rounded and the result converted to an integer.
Text is converted to a text:sequence pair and saved to a CAT file, and the sequence is converted to the TIF.
}
