% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_multisurface_gdb.R
\name{check_multisurface_gdb}
\alias{check_multisurface_gdb}
\title{Check for MULTISURFACE or other unsupported geometry types in a File Geodatabase.}
\usage{
check_multisurface_gdb(gdb_path, layer_name, nrow = 1000)
}
\arguments{
\item{gdb_path}{Character. The full file path to the File Geodatabase (.gdb).}

\item{nrow}{Logical. Returns `TRUE` if unsupported geometry types like `MULTISURFACE` are detected; `FALSE` otherwise.}

\item{Character.}{The name of the feature class (layer) within the geodatabase to check.}
}
\description{
Check for MULTISURFACE or other unsupported geometry types in a File Geodatabase.
}
\examples{
in_gdb <- "E:/data/indata.gdb"
in_fc  <- "parks"

check_geom <- check_multisurface_gdb(
  gdb_path = in_gdb,
  layer_name = in_fc,
  nrow = 1000
)

if (!check_geom) {
  message(glue::glue("No MULTISURFACE geometry found in {in_gdb}/{in_fc}"))
} else {
  warning(glue::glue("MULTISURFACE geometry detected in {in_gdb}/{in_fc}. Data may not import correctly."))
}
}
