% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-mask-tif.R
\name{create_mask_tif}
\alias{create_mask_tif}
\title{Creates a raster mask from a vector input}
\usage{
create_mask_tif(
  in_src,
  field,
  template,
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  in_lyr = NULL,
  out_tif_path = NULL,
  out_tif_name = NULL,
  datatype = "INT4S"
)
}
\arguments{
\item{in_src}{path of gdb or geopackage}

\item{field}{field to be rasterized}

\item{template}{path of template raster}

\item{crop_extent}{OPTIONAL - extent of output raster c(<xmin>,<ymax>,<ymin>,<xmax>) default c(273287.5,1870587.5,367787.5,1735787.5)}

\item{in_lyr}{OPTIONAL - input data layer e.g. 'veg_comp_poly'}

\item{out_tif_path}{path of out TIF}

\item{out_tif_name}{name of out TIF}

\item{datatype}{defaults to Int64}
}
\value{
output tif name
}
\description{
Creates a raster mask from a vector input
}
\examples{
coming soon
}
