% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMaskFromVect.R
\name{createMaskFromVect}
\alias{createMaskFromVect}
\title{Creates a raster mask from a vector input}
\usage{
createMaskFromVect(
  inSrc,
  field,
  template,
  cropExtent = NULL,
  inlyr = NULL,
  outTifpath = NULL,
  outTifname = NULL,
  datatype = "INT4S"
)
}
\arguments{
\item{inSrc}{path of gdb or geopackage}

\item{field}{field to be rasterized}

\item{template}{path of template raster}

\item{cropExtent}{OPTIONAL - extent of output raster <xmin>,<ymax>,<ymin>,<xmax> default c(273287.5,1870587.5,367787.5,1735787.5)}

\item{inlyr}{OPTIONAL - input data layer e.g. 'veg_comp_poly'}

\item{outTifpath}{path of out TIF defualt D:/Projects/provDataProject}

\item{outTifname}{name of out TIF}

\item{datatype}{default Int64}
}
\value{
output tif name
}
\description{
Creates a raster mask from a vector input
}
\examples{
coming soon
}
