% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-to-pg.R
\name{df_to_pg}
\alias{df_to_pg}
\title{Write a PG Table from Dataframe}
\usage{
df_to_pg(pg_tbl, in_df, pg_conn_param, overwrite = TRUE, append = FALSE)
}
\arguments{
\item{pg_tbl}{Name of output PG Table}

\item{in_df}{input dataframe}

\item{pg_conn_param}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}

\item{overwrite}{overwrite output T or F}

\item{append}{append output T or F}
}
\value{
no return
}
\description{
Write a PG Table from Dataframe
}
\examples{
df_to_pg(pg_tbl = 'pgTableName', in_df = in_df, pg_conn_params = get_pg_conn_list(), overwrite = 'TRUE', append = 'FALSE')
}
