% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfColNames2lower.R
\name{dfColNames2lower}
\alias{dfColNames2lower}
\title{Format colnames for PG (i.e. lower case and replace '.' with '_')}
\usage{
dfColNames2lower(df)
}
\arguments{
\item{df}{input dataframe}
}
\value{
no return
}
\description{
Format colnames for PG (i.e. lower case and replace '.' with '_')
}
\examples{
coming soon
}
