% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdw-to-geom-tbl.R
\name{fdw_to_geom_tbl}
\alias{fdw_to_geom_tbl}
\title{Create a PG spatial table from Foreign data wrapper table}
\usage{
fdw_to_geom_tbl(
  ora_tbl,
  out_tbl,
  pk,
  out_schema,
  ora_conn_param,
  fdw_schema = "load"
)
}
\arguments{
\item{ora_tbl}{coming soon}

\item{out_tbl}{coming soon}

\item{pk}{coming soon}

\item{ora_conn_param}{coming soon}

\item{schema}{coming soon}
}
\value{
None
}
\description{
Create a PG spatial table from Foreign data wrapper table
}
\examples{
coming soon
}
