% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdw-to-tbl.R
\name{fdw_to_tbl}
\alias{fdw_to_tbl}
\title{Create a PG non spatial table from Foreign data wrapper table}
\usage{
fdw_to_tbl(
  src_tbl,
  dst_tbl,
  pk,
  dst_schema,
  pg_conn_param,
  tbl_comment,
  src_schema = "load",
  flds_to_keep = NULL,
  where = ""
)
}
\arguments{
\item{src_tbl}{coming soon}

\item{dst_tbl}{coming soon}

\item{pk}{coming soon}

\item{dst_schema}{coming soon}

\item{pg_conn_param}{coming soon}

\item{tbl_comment}{coming soon}

\item{src_schema}{optional}

\item{flds_to_keep}{optional}

\item{where}{optional}
}
\value{
None
}
\description{
Create a PG non spatial table from Foreign data wrapper table
}
\examples{
coming soon
}
