% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdwTbl2PGnoSpatial.R
\name{fdwTbl2PGnoSpatial}
\alias{fdwTbl2PGnoSpatial}
\title{Create a PG non spatial table from Foreign data wrapper table}
\usage{
fdwTbl2PGnoSpatial(
  src_tbl,
  dst_tbl,
  pk,
  dst_schema,
  pg_conn_param,
  tbl_comment,
  src_schema = "load",
  flds_to_keep = NULL,
  where = ""
)
}
\arguments{
\item{src_tbl}{coming soon}

\item{dst_tbl}{coming soon}

\item{pk}{coming soon}

\item{pg_conn_param}{coming soon}

\item{tbl_comment}{}

\item{src_schema}{coming soon}

\item{flds_to_keep}{optional}

\item{schema}{coming soon}
}
\value{
None
}
\description{
Create a PG non spatial table from Foreign data wrapper table
}
\examples{
coming soon
}
