% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdwTbl2PGSpatial.R
\name{fdwTbl2PGSpatial}
\alias{fdwTbl2PGSpatial}
\title{Create a PG spatial table from Foreign data wrapper table}
\usage{
fdwTbl2PGSpatial(
  oratable,
  outTblName,
  pk,
  outSchema,
  connList,
  fdwSchema = "load"
)
}
\arguments{
\item{oratable}{coming soon}

\item{outTblName}{coming soon}

\item{pk}{coming soon}

\item{connList}{coming soon}

\item{schema}{coming soon}
}
\value{
None
}
\description{
Create a PG spatial table from Foreign data wrapper table
}
\examples{
coming soon
}
