% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalCropRaster.R
\name{gdalCropRaster}
\alias{gdalCropRaster}
\title{Use gdal to crop Raster to specified extent}
\usage{
gdalCropRaster(inTif, cropExtent, outTif, res = NULL)
}
\arguments{
\item{inTif}{path and name to TIF}

\item{cropExtent}{vector with extents in the following order  <xmin>,<ymax>,<ymin>,<xmax>}

\item{outTif}{output tif path and name}

\item{res}{resolution default is 100}
}
\value{
no return
}
\description{
Use gdal to crop Raster to specified extent
}
\examples{
gdalCropRaster('myTif.tif',extVect,'myTifCropped.tif')
}
