% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-square-from-point.R
\name{geom_square_from_point}
\alias{geom_square_from_point}
\title{Add a polyon geometry field (1 hectare squares) to
a PostGres table and populate using a point geometry field.}
\usage{
geom_square_from_point(
  pg_conn_param = get_pg_conn_list(),
  pg_geomTable,
  pg_exisiting_geomPoint,
  pg_add_geomName
)
}
\arguments{
\item{pg_conn_param}{named (optional): list - connection parameters (default returned by dadmtools::get_pg_conn_list)}

\item{pg_geomTable}{(required): string - PG table with point geometry}

\item{pg_exisiting_geomPoint}{(required): Name of point geometry field}

\item{pg_add_geomName}{(required): string - name of new polygon geometry field}
}
\description{
Add a polyon geometry field (1 hectare squares) to
a PostGres table and populate using a point geometry field.
}
