% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gr-skey-extent.R
\name{get_gr_skey_extent}
\alias{get_gr_skey_extent}
\title{Get extent from vector dataset that snaps to template raster}
\usage{
get_gr_skey_extent(
  dsn,
  layer,
  template_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif"
)
}
\arguments{
\item{dsn}{coming soon}

\item{layer}{coming soon}

\item{template_tif}{coming soon}
}
\value{
extent vector (i.e. c(xmin,xmax,ymin,ymax))
}
\description{
Get extent from vector dataset that snaps to template raster
}
\examples{
coming soon
}
