% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ora_conn_list.R
\name{get_ora_conn_list}
\alias{get_ora_conn_list}
\title{Get named list of  oracle connection parameters}
\usage{
get_ora_conn_list(
  host = NULL,
  user = NULL,
  serviceName = NULL,
  server = NULL,
  password = NULL,
  port = "1521"
)
}
\arguments{
\item{host}{oracle database host. Default is oracle keyring value.}

\item{user}{oracle database user. Default is oracle keyring value.}

\item{serviceName}{oracle service name. Default is oracle keyring value.}

\item{password}{pg database password. Default is oracle keyring value.}

\item{port}{pg database port. Default '1521'.}
}
\value{
returns named list of connection parameters
}
\description{
Get named list of  oracle connection parameters
}
\examples{
get_ora_conn_list(schema = 'whse', serviceName = 'test')
}
