% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-oracle-conn-string.R
\name{get_oracle_conn_string}
\alias{get_oracle_conn_string}
\title{Get named oracle connection string}
\usage{
get_oracle_conn_string(
  in_tbl_name,
  schema = NULL,
  host = NULL,
  user = NULL,
  service_name = NULL,
  password = NULL,
  port = "1521"
)
}
\arguments{
\item{in_tbl_name}{Oracle Table Name}

\item{schema}{the database schema default is NULL and implies the shcma name will be in the the input table name}

\item{host}{database host. Default is oracle keyring value.}

\item{user}{database user. Default is oracle keyring value.}

\item{service_name}{database service name. Default is oracle keyring value.}

\item{password}{database password. Default is oracle keyring value.}

\item{port}{database port. Default '1521'.}
}
\value{
returns named list of connection parameters
}
\description{
Get named oracle connection string
}
\examples{
get_oracle_conn_string(schema = 'whse', dbname = 'test')
}
