% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pg_conn_list.R
\name{get_pg_conn_list}
\alias{get_pg_conn_list}
\title{Get named list of  postgres connection parameters}
\usage{
get_pg_conn_list(
  driver = NULL,
  host = NULL,
  user = NULL,
  dbname = NULL,
  password = NULL,
  port = "5432"
)
}
\arguments{
\item{driver}{the database driver. Default is keyring value. E.g. dbDriver("PostgreSQL")}

\item{host}{pg database host. Default is localpsql keyring value.}

\item{user}{pg database user. Default is localpsql keyring value.}

\item{dbname}{pg database name. Default is localpsql keyring value.}

\item{password}{pg database password. Default is localpsql keyring value.}

\item{port}{pg database port. Default '5432'.}
}
\value{
returns named list of connection parameters
}
\description{
Get named list of  postgres connection parameters
}
\examples{
get_pg_conn_list(dbname = 'postgres', keyring::key_get('dbpass', keyring = 'localpsql'))
}
