% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pg-geom-name.R
\name{get_pg_geom_name}
\alias{get_pg_geom_name}
\title{Get name of geometry name is PG}
\usage{
get_pg_geom_name(schema, tablename, pg_conn_param)
}
\arguments{
\item{schema}{Name of table's schema}

\item{tablename}{Name of pg table}

\item{pg_conn_param}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}
}
\value{
returns field name of geometry field
}
\description{
Get name of geometry name is PG
}
\examples{
coming soon
}
