% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pg-tbl-by-pattern.R
\name{get_pg_tbl_by_pattern}
\alias{get_pg_tbl_by_pattern}
\title{Query PG and return list of table names that are date based (named)}
\usage{
get_pg_tbl_by_pattern(pattern, schema, pg_conn_param, depth = 2)
}
\arguments{
\item{pattern}{An text string for the table pattern to look for date named related tables}

\item{schema}{The schema to search in}

\item{pg_conn_param}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}

\item{depth}{The number of matches with dates to return. default is 2}
}
\value{
a dataframe of the list of matching tables
}
\description{
Query PG and return list of table names that are date based (named)
}
\examples{
get_pg_tbl_by_pattern('techpanel_scenarios','localhost','postgres','myDB','mypassword',5432,'prod')
}
