% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeomNamePG.R
\name{getGeomNamePG}
\alias{getGeomNamePG}
\title{Get name of geometry name is PG}
\usage{
getDatePatternTableList(pattern, connList, depth)
}
\arguments{
\item{pattern}{String pattern to match for date named tables}

\item{connList}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}

\item{depth}{The number of date patterned matches to return}
}
\value{
Data Frame of the list of matching table names 
}
\description{
Return a dataframe list of the date named tables matching a pattern
}
\examples{
coming soon
}
