% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFDWtblSpSQL.R
\name{getFDWtblSpSQL}
\alias{getFDWtblSpSQL}
\title{Returns a SQL statement including the spatial field and primary key field of a Foreign Data Wrapper table.  Can be used in gdal rasterize statement}
\usage{
getFDWtblSpSQL(
  dst_table_name,
  dst_schema_name,
  oratable,
  pk,
  connList,
  fdwSchema = "load",
  where = NULL
)
}
\arguments{
\item{oratable}{coming soon}

\item{pk}{coming soon}

\item{connList}{coming soon}

\item{fdwSchema}{coming soon}

\item{where}{coming soon}
}
\value{
sql string
}
\description{
Returns a SQL statement including the spatial field and primary key field of a Foreign Data Wrapper table.  Can be used in gdal rasterize statement
}
\examples{
coming soon
}
