% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeomNamePG.R
\name{getGeomNamePG}
\alias{getGeomNamePG}
\title{Get name of geometry name is PG}
\usage{
getGeomNamePG(schema, tablename, connList)
}
\arguments{
\item{schema}{Name of table's schema}

\item{tablename}{Name of pg table}

\item{connList}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}
}
\value{
returns field name of geometry field
}
\description{
Get name of geometry name is PG
}
\examples{
coming soon
}
