% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrskeyExtent.R
\name{getGrskeyExtent}
\alias{getGrskeyExtent}
\title{Get extent from vector dataset that snaps to template raster}
\usage{
getGrskeyExtent(
  dsn,
  layer,
  templateRaster =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif"
)
}
\arguments{
\item{dsn}{coming soon}

\item{layer}{coming soon}

\item{templateRaster}{coming soon}
}
\value{
extent vector (i.e. c(xmin,xmax,ymin,ymax))
}
\description{
Get extent from vector dataset that snaps to template raster
}
\examples{
coming soon
}
