% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableQueryPG.R
\name{getTableQueryPG}
\alias{getTableQueryPG}
\title{Query PG and return results to a dataframe}
\usage{
getTableQueryPG(sql, connList)
}
\arguments{
\item{sql}{A sql string}

\item{connList}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}
}
\value{
a dataframe of the returned query
}
\description{
Query PG and return results to a dataframe
}
\examples{
getTableQueryPG('select * from table','localhost','postgres','myDB','mypassword',5432,'prod')
}
