% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_skey_tif_2_pg_geom.R
\name{gr_skey_tif_2_pg_geom}
\alias{gr_skey_tif_2_pg_geom}
\title{Import Grskey TIF 2 pg table}
\usage{
gr_skey_tif_2_pg_geom(
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Lands_and_Islandsincluded.tif",
  cropExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  outCropTifName = "D:\\\\Projects\\\\provDataProject\\\\gr_skey_cropped.tif",
  connList = faibDataManagement::get_pg_conn_list()
)
}
\arguments{
\item{grskeyTIF}{coming soon}

\item{maskTif}{coming soon}

\item{cropExtent}{coming soon}

\item{outCropTifName}{coming soon}

\item{connList}{coming soon}
}
\value{
coming soon
}
\description{
Import Grskey TIF 2 pg table
}
\examples{
coming soon
}
