% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gr-skey-tif-to-pg-rast.R
\name{import_gr_skey_tif_to_pg_rast}
\alias{import_gr_skey_tif_to_pg_rast}
\title{Import Grskey TIF 2 pg table}
\usage{
import_gr_skey_tif_to_pg_rast(
  out_crop_tif_name,
  template_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  mask_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  pg_conn_param = dadmtools::get_pg_conn_list(),
  dst_tbl = "whse.all_bc_gr_skey",
  rast_sch = "raster",
  pg_rast_name = "grskey_bc_land",
  geom_type = "Centroid"
)
}
\arguments{
\item{out_crop_tif_name}{filename of the written output tif}

\item{template_tif}{The file path to the gr_skey template geotiff, defaults to  "S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\bc_01ha_gr_skey.tif"}

\item{mask_tif}{The file path to the geotiff to be used as a mask, defaults to "S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\BC_Boundary_Terrestrial.tif"}

\item{crop_extent}{Raster crop extent, list of c(ymin, ymax, xmin, xmax) in EPSG:3005, defaults to c(273287.5,1870587.5,367787.5,1735787.5)}

\item{pg_conn_param}{Keyring object of Postgres credentials, defaults to dadmtools::get_pg_conn_list()}

\item{dst_tbl}{Destination table of the imported gr_skey_tbl table (format: schema_name.table_name), defaults to "whse.all_bc_gr_skey"}

\item{rast_sch}{Destination schema, defaults to "raster"}

\item{pg_rast_name}{Destination pg raster name, defaults to "grskey_bc_land"}

\item{geom_type}{Available options are: Centroid or Polygon. Defaults to "Centroid"}
}
\value{
coming soon
}
\description{
Import Grskey TIF 2 pg table
}
\examples{
coming soon
}
