% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-to-pg-gr-skey-overlap.R
\name{import_to_pg_gr_skey_overlap}
\alias{import_to_pg_gr_skey_overlap}
\title{Import data with spatial overlaps into Postgres as gr_skey table}
\usage{
import_to_pg_gr_skey_overlap(
  inSrc,
  inLyr,
  groupFld,
  outPGTblName,
  schema = "whse",
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  pg_conn_param = dadmtools::get_pg_conn_list(),
  terraExt = terra::ext(c(273287.5, 1870587.5, 367787.5, 1735787.5))
)
}
\arguments{
\item{inSrc}{input source}

\item{inLyr}{input layer}

\item{groupFld}{coming soon}

\item{outPGTblName}{output pg table name}

\item{schema}{Defaults to 'whse'}

\item{grskeyTIF}{Defaults to 'S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\bc_01ha_gr_skey.tif'}

\item{maskTif}{Defaults to 'S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\BC_Boundary_Terrestrial.tif'}

\item{pg_conn_param}{Defaults to get_pg_conn_list()}

\item{terraExt}{Defaults to terra::ext(c(273287.5,1870587.5,367787.5,1735787.5))}
}
\value{
no return
}
\description{
Import data with spatial overlaps into Postgres as gr_skey table
}
\examples{
coming soon

}
\keyword{internal}
