% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowercase-df-colnames.R
\name{lowercase_df_colnames}
\alias{lowercase_df_colnames}
\title{Format colnames for PG (i.e. lower case and replace '.' with '_')}
\usage{
lowercase_df_colnames(df)
}
\arguments{
\item{df}{input dataframe}
}
\value{
no return
}
\description{
Format colnames for PG (i.e. lower case and replace '.' with '_')
}
\examples{
coming soon
}
