% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr-to-tbl.R
\name{ogr_to_tbl}
\alias{ogr_to_tbl}
\title{Write non spatial tables to PG using gdal on the machine.  Creates index on primary key.  Use for large files that would take too long to import into R}
\usage{
ogr_to_tbl(
  src,
  dst_tbl,
  pg_conn_param,
  lyr = NULL,
  pk = NULL,
  select = NULL,
  where = NULL,
  dst_schema = NULL,
  tbl_comment = NULL
)
}
\arguments{
\item{src}{path of input dataset e.g. c:/test.gdb}

\item{dst_tbl}{Name of output PG table}

\item{lyr}{layer of input dataser e.g. roadsFC}

\item{pk}{primary of table e.g. objectid}

\item{select}{fields to keep in table e.g. "objectid,feature_id,segmentName"}

\item{dst_schema}{default database schema}

\item{dbName}{Name of PG database}

\item{where_clause}{sql where clause e.g. segment = 'highway2'}

\item{useGdal}{Set to TRUE if GDAL is installed on machine e.g. TRUE}
}
\value{
no return
}
\description{
Write non spatial tables to PG using gdal on the machine.  Creates index on primary key.  Use for large files that would take too long to import into R
}
\examples{
coming soon
}
