% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_rast_2_centroid.R
\name{pg_rast_2_centroid}
\alias{pg_rast_2_centroid}
\title{Converts postgres raster to  posgres points table}
\usage{
pg_rast_2_centroid(
  inRas,
  outLyrName,
  outField,
  connList = faibDataManagement::get_pg_conn_list()
)
}
\arguments{
\item{inRas}{coming soon}

\item{outLyrName}{coming soon}

\item{outField}{coming soon}
}
\value{
coming soon
}
\description{
Converts postgres raster to  posgres points table
}
\examples{
coming soon
}
