% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-rast-to-centroid.R
\name{pg_rast_to_centroid}
\alias{pg_rast_to_centroid}
\title{Converts postgres raster to postgres points table}
\usage{
pg_rast_to_centroid(
  in_ras,
  out_lyr_name,
  out_field,
  pg_conn_param = dadmtools::get_pg_conn_list()
)
}
\arguments{
\item{in_ras}{coming soon}

\item{out_lyr_name}{coming soon}

\item{out_field}{coming soon}

\item{pg_conn_param}{Keyring object of Postgres credentials}
}
\value{
coming soon
}
\description{
Converts postgres raster to postgres points table
}
\examples{
coming soon
}
