% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-rast-to-tbl.R
\name{pg_rast_to_tbl}
\alias{pg_rast_to_tbl}
\title{converts a PG raster into a non spatial pg table}
\usage{
pg_rast_to_tbl(out_lyr_name, in_ras, out_field, template_ras, pg_conn_param)
}
\arguments{
\item{out_lyr_name}{name of output non-spatial table}

\item{in_ras}{input PG Raster}

\item{out_field}{name of the value field in output table}

\item{pg_conn_param}{coming soon}

\item{templateRas}{template pg raster with the same the resolution and extent of the input raster}
}
\value{
nothing is returned
}
\description{
converts a PG raster into a non spatial pg table
}
\examples{
coming soon
}
