% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-to-gdb.R
\name{pg_to_gdb}
\alias{pg_to_gdb}
\title{Convert a Postgres table to GDB}
\usage{
pg_to_gdb(
  pg_conn_param = dadmtools::get_pg_conn_list(),
  fields_to_gdb = NULL,
  pg_att_table,
  pg_attskey_table = NULL,
  pg_skgeom_table = NULL,
  geom_field,
  dst_ha_field = "fid_ha",
  key_skgeo_tbl = "gr_skey",
  key_attskey_tbl = "pgid",
  query = NULL,
  dst_gdb_name,
  out_gdb_path
)
}
\arguments{
\item{pg_conn_param}{named (optional): list - connection parameters. Default is  dadmtools::get_pg_conn_list().}

\item{fields_to_gdb}{(optional): string - comma separated list of fields to export. Default is all fields}

\item{pg_att_table}{(required): string - name of Postgres table with fields to export to GDB}

\item{pg_attskey_table}{(optional): string - name of Postgres table with join field to `pg_att_table` and table with geometry.}

\item{pg_skgeom_table}{(optional): string - name of Postgres table with geometry and join field to `pg_att_table` or `pg_attskey_table`.}

\item{geom_field}{(required): string - name of geometry field}

\item{dst_ha_field}{(optional): string - name of hectares field created in destination GDB. Default is `fid_ha`.}

\item{key_skgeo_tbl}{(optional): string - name of key field (e.g. gr_skey) to join geometry table (`pg_skgeom_table`) with either `pg_att_table` or `pg_attskey_table`.}

\item{key_attskey_tbl}{(optional): string - name of key field (e.g. pgid) to join attribute table (`pg_att_table`) with join table (`pg_attskey_table`).}

\item{query}{(optional): string - SQL query to filter results (e.g. thlb_fact > 0). SQL statement placed after `where`. Default is all rows returned.}

\item{dst_gdb_name}{(required): string - name of destination GDB being created.}

\item{out_gdb_path}{(required): string - path to directory where GDB will be saved to.}
}
\description{
Convert a Postgres table to GDB
}
