% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-to-seles-tif.R
\name{pg_to_seles_tif}
\alias{pg_to_seles_tif}
\title{Rasterize Field in Postgres Table (dataypes: integer, decimal, text)
Seles only uses integer rasters.
If the field is an integer, a TIF file is created with the integer value
If the field is a text field, a CAT file will be created, a
value - integer pairing. The TIF file will be created from the integer.
The CAT file is used in Seles to display the legend.
If the filed file is a decimal, it will be multiplied by the parameter d_multiplier, then converted
to an integer. A TIF file will be created from the integer.}
\usage{
pg_to_seles_tif(
  pg_conn_param,
  field_to_tif,
  pg_att_table,
  pg_attskey_table = NULL,
  pg_skgeom_table = NULL,
  geom_field,
  key_skgeo_tbl = NULL,
  key_attskey_tbl = NULL,
  template_tif,
  out_tif_path,
  dst_tif_name,
  out_cat_path = NULL,
  d_multiplier = 1000,
  query = NULL,
  gdb,
  bnd
)
}
\arguments{
\item{pg_conn_param}{named (required): list - connection parameters (returned by dadmtools::get_pg_conn_list)}

\item{field_to_tif}{(required): string - field name in pg_table_with_geom to convert to raster}

\item{pg_att_table}{(required): string - name of Postgres schema.table with field to convert. Optionally can have geometry field}

\item{pg_attskey_table}{(optional): string - name of Postgres schema.table with key to geom and attribute tables (e.g. gr_skey or ogc_fid),   required if geometry not in attribute table}

\item{pg_skgeom_table}{(optional): string - name of Postgres schema.table with geometry, if geometry not in attribute table}

\item{geom_field}{(required): string - name of geometry field in pg_table_with_geom}

\item{key_skgeo_tbl}{(optional): string - name of join field to join pg_skey_table and pg_geom_table (e.g. gr_skey), required if joining 3 tables.}

\item{key_attskey_tbl}{(optional): string - name of join field to join pg_att_table and pg_skey_table (e.g. pgid), required if joining 2 or 3 tables.}

\item{template_tif}{(required): string - file path to template tif (e.g. skey.tif)}

\item{out_tif_path}{(required): string - path to output rasters directory}

\item{dst_tif_name}{(required): string - name of output tif saved to out_tif_path}

\item{out_cat_path}{(optional): string - path to output cat files directory. Only used for text field types. Default NULL.}

\item{query}{(optional): string SQL where clause used to filter input dataset, e.g. "thlb_fact > 0", "own is not null". Default NULL (no query applied).}

\item{d_mulitiplier}{(optional): integer multiplier to use for decimal datatypes. Only used for decimal data types. Default 1000}
}
\description{
Rasterize Field in Postgres Table (dataypes: integer, decimal, text)
Seles only uses integer rasters.
If the field is an integer, a TIF file is created with the integer value
If the field is a text field, a CAT file will be created, a
value - integer pairing. The TIF file will be created from the integer.
The CAT file is used in Seles to display the legend.
If the filed file is a decimal, it will be multiplied by the parameter d_multiplier, then converted
to an integer. A TIF file will be created from the integer.
}
