% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-to-shapefile.R
\name{pg_to_shapefile}
\alias{pg_to_shapefile}
\title{Create a shapefile from fields in a Postgres Table}
\usage{
pg_to_shapefile(
  pg_conn_param = dadmtools::get_pg_conn_list(),
  fields_to_shape,
  pg_att_table,
  pg_attskey_table = NULL,
  pg_skgeom_table = NULL,
  geom_field,
  key_skgeo_tbl = NULL,
  key_attskey_tbl = NULL,
  out_shapefile_path,
  dst_shapefile_name,
  query
)
}
\arguments{
\item{pg_conn_param}{named (optional): list - connection parameters (default is returned by dadmtools::get_pg_conn_list)}

\item{fields_to_shape}{(required): string - field names to convert to shapefile - if more than one,must be separated by a comma}

\item{pg_att_table}{(required): string - name of Postgres schema.table with field to convert. Optionally can have geometry field}

\item{pg_attskey_table}{(optional): string - name of Postgres schema.table with key to geom and attribute tables (e.g. gr_skey or ogc_fid),   required if geometry not in attribute table}

\item{pg_skgeom_table}{(optional): string - name of Postgres schema.table with geometry, if geometry not in attribute table}

\item{geom_field}{(required): string - name of geometry field in pg_table_with_geom}

\item{key_skgeo_tbl}{(optional): string - name of join field to join pg_skey_table and pg_geom_table (e.g. gr_skey), required if joining 3 tables.}

\item{key_attskey_tbl}{(optional): string - name of join field to join pg_att_table and pg_skey_table (e.g. pgid), required if joining 2 or 3 tables.}

\item{out_shapefile_path}{(required): string - path to output shapefile directory}

\item{dst_shapefile_name}{(required): string - name of output shapefile saved to out_shapefile_path}

\item{query}{(optional): string SQL where clause used to filter input dataset, e.g. "thlb_fact > 0", "own is not null". Default NULL (no query applied).}
}
\description{
Create a shapefile from fields in a Postgres Table
}
