% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgRas2rows.R
\name{pgRas2rows}
\alias{pgRas2rows}
\title{converts a PG raster into a non spatial pg table}
\usage{
pgRas2rows(outLyrName, inRas, outField, templateRas, connList)
}
\arguments{
\item{outLyrName}{name of output non-spatial table}

\item{inRas}{input PG Raster}

\item{outField}{name of the value field in output table}

\item{templateRas}{template pg raster with the same the rsolution znd extent of the input raster}
}
\value{
nothing is returned
}
\description{
converts a PG raster into a non spatial pg table
}
\examples{
coming soon
}
