% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize-gdal.R
\name{rasterize_gdal}
\alias{rasterize_gdal}
\title{Rasterize spatial data using Gdal on setup on local machine}
\usage{
rasterize_gdal(
  in_lyr,
  field,
  out_tif_path,
  out_tif_name,
  src_path = NULL,
  pg_conn_param = NULL,
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  nodata = 0,
  where = NULL,
  datatype = "UInt32"
)
}
\arguments{
\item{in_lyr}{input data layer e.g. 'veg_comp_poly'}

\item{field}{field to be rasterized}

\item{out_tif_path}{path of out TIF defualt D:/Projects/provDataProject}

\item{out_tif_name}{name of out TIF}

\item{src_path}{path pf gdb or geopackage}

\item{pg_conn_param}{Named list with the following pg connection parameters Driver,host,user,dbname,password,port,schema}

\item{crop_extent}{vector with extents in the following order  <xmin>,<ymax>,<ymin>,<xmax> default c(273287.5,1870587.5,367787.5,1735787.5)}

\item{nodata}{no data value of out tif default is 0}

\item{where}{default NULL}

\item{datatype}{default Int64}
}
\value{
output tif name
}
\description{
Rasterize spatial data using Gdal on setup on local machine
}
\examples{
comming soon
}
