% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize-sf.R
\name{rasterize_sf}
\alias{rasterize_sf}
\title{Rasterize spatial data using SF.  Accepts postgis, gdb and geopackage}
\usage{
rasterize_sf(
  in_lyr,
  field,
  out_tif_dir = "D:\\\\Projects\\\\provDataProject",
  out_tif_name,
  in_src = NULL,
  pg_conn_param = NULL,
  vec_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  nodata = 0,
  where = NULL,
  datatype = "Int32L"
)
}
\arguments{
\item{in_lyr}{input data layer e.g. 'veg_comp_poly'}

\item{field}{field to be rasterized}

\item{out_tif_dir}{path of out TIF defualt D:/Projects/provDataProject}

\item{out_tif_name}{name of out TIF}

\item{in_src}{path pf gdb or geopackage}

\item{pg_conn_param}{Named list with the following pg connection parameters Driver,host,user,dbname,password,port,schema}

\item{vec_extent}{vector with extents in the following order  <xmin>,<ymax>,<ymin>,<xmax> default c(273287.5,1870587.5,367787.5,1735787.5)}

\item{nodata}{no data value of out tif defualt is 0}
}
\value{
output tif name
}
\description{
Rasterize spatial data using SF.  Accepts postgis, gdb and geopackage
}
\examples{
coming soon


}
