% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize-terra.R
\name{rasterize_terra}
\alias{rasterize_terra}
\title{Rasterize spatial data using Terra}
\usage{
rasterize_terra(
  src_sf,
  field,
  template_tif,
  crop_extent = NULL,
  src_lyr = NULL,
  out_tif_path = NULL,
  out_tif_name = NULL,
  datatype = "INT4S",
  nodata = 0
)
}
\arguments{
\item{src_sf}{Simple feature object}

\item{field}{field to be rasterized}

\item{crop_extent}{OPTIONAL - extent of output raster <xmin>,<ymax>,<ymin>,<xmax> default c(273287.5,1870587.5,367787.5,1735787.5)}

\item{src_lyr}{OPTIONAL - input data layer e.g. 'veg_comp_poly'}

\item{out_tif_path}{path of out TIF defualt D:/Projects/provDataProject}

\item{out_tif_name}{name of out TIF}

\item{datatype}{default Int64}

\item{template}{path of template raster}
}
\value{
output tif name
}
\description{
Rasterize spatial data using Terra
}
\examples{
coming soon
}
