% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeSF.R
\name{rasterizeSF}
\alias{rasterizeSF}
\title{Rasterize spatial data using SF.  Accepts postgis, gdb and geopackage}
\usage{
rasterizeSF(
  inlyr,
  field,
  outTifpath = "D:\\\\Projects\\\\provDataProject",
  outTifname,
  inSrc = NULL,
  pgConnList = NULL,
  vecExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  nodata = 0,
  where = NULL,
  datatype = "Int32L"
)
}
\arguments{
\item{inlyr}{input data layer e.g. 'veg_comp_poly'}

\item{field}{field to be rasterized}

\item{outTifpath}{path of out TIF defualt D:/Projects/provDataProject}

\item{outTifname}{name of out TIF}

\item{inSrc}{path pf gdb or geopackage}

\item{pgConnList}{Named list with the following pg connection parameters Driver,host,user,dbname,password,port,schema}

\item{vecExtent}{vector with extents in the following order  <xmin>,<ymax>,<ymin>,<xmax> default c(273287.5,1870587.5,367787.5,1735787.5)}

\item{nodata}{no data value of out tif defualt is 0}
}
\value{
output tif name
}
\description{
Rasterize spatial data using SF.  Accepts postgis, gdb and geopackage
}
\examples{
coming soon


}
