% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeTerra.R
\name{rasterizeTerra}
\alias{rasterizeTerra}
\title{Rasterize spatial data using Terra}
\usage{
rasterizeTerra(
  inSrc,
  field,
  template,
  cropExtent = NULL,
  inlyr = NULL,
  outTifpath = NULL,
  outTifname = NULL,
  datatype = "INT4S",
  nodata = 0
)
}
\arguments{
\item{inSrc}{path of gdb or geopackage}

\item{field}{field to be rasterized}

\item{template}{path of template raster}

\item{cropExtent}{OPTIONAL - extent of output raster <xmin>,<ymax>,<ymin>,<xmax> default c(273287.5,1870587.5,367787.5,1735787.5)}

\item{inlyr}{OPTIONAL - input data layer e.g. 'veg_comp_poly'}

\item{outTifpath}{path of out TIF defualt D:/Projects/provDataProject}

\item{outTifname}{name of out TIF}

\item{datatype}{default Int64}
}
\value{
output tif name
}
\description{
Rasterize spatial data using Terra
}
\examples{
coming soon
}
