% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeWithGdal.R
\name{rasterizeWithGdal}
\alias{rasterizeWithGdal}
\title{Rasterize spatial data using Gdal on setup on local machine}
\usage{
rasterizeWithGdal(pk, src, lyr, extent, where_clause = NULL, outName = NULL)
}
\arguments{
\item{pk}{field name of whose values will be used in the raster (must be a numric type field e.g. integers, float etc)}

\item{src}{path of input dataset (e.g D:/data/data.gdb)}

\item{lyr}{input data layer e.g. 'veg_comp_poly'}

\item{where_clause}{sql where use to filter data}

\item{outName}{optional output name of tif default is the lyr parameter concatenated with the pk parameter}

\item{cropExtent}{vector with extents in the following order  <xmin>,<ymax>,<ymin>,<xmax>}
}
\value{
output tif name
}
\description{
Rasterize spatial data using Gdal on setup on local machine
}
\examples{
comming soon
}
