% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-sql-psql.R
\name{run_sql_psql}
\alias{run_sql_psql}
\title{Run a external sql file}
\usage{
run_sql_psql(sql_var, sql_file, pg_db, host = NULL)
}
\arguments{
\item{sql_var}{vector of variables to use in sql file e.g. c("schema='dbName',xwalkprev='inFC')}

\item{sql_file}{path and file name of sql file}

\item{pg_db}{pg database name}

\item{host}{host of db e.g. 'localhost'}
}
\value{
no return
}
\description{
Run a external sql file
}
\examples{
sql_var <- c("schema"='dbName',xwalkprev='inFC') \cr
sql_file <- C:/sql/file.sql \cr
run_sql_psql(sql_var,sql_file,pg_db)
}
