% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-sql-r.R
\name{run_sql_r}
\alias{run_sql_r}
\title{Send sql statement to PG}
\usage{
run_sql_r(sql, pg_conn_param)
}
\arguments{
\item{sql}{A sql string}

\item{pg_conn_param}{Named list with the following connection parameters Driver,host,user,dbname,password,port,schema}
}
\value{
no return
}
\description{
Send sql statement to PG
}
\examples{
run_sql_r('Create table tester as select * from test1','postgres','myDB','mypassword',5432,'prod')
}
