% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSQL.R
\name{runSQL}
\alias{runSQL}
\title{Run a external sql file}
\usage{
runSQL(sqlVar, sqlFile, pgDB, host = NULL)
}
\arguments{
\item{sqlVar}{vector of variables to use in sql file e.g. c("schema='dbName',xwalkprev='inFC')}

\item{sqlFile}{path and file name of sql file}

\item{pgDB}{pg database name}

\item{host}{host of db e.g. 'localhost'}
}
\value{
no return
}
\description{
Run a external sql file
}
\examples{
sqlVar <- c("schema"='dbName',xwalkprev='inFC') \cr
sqlFile <- C:/sql/file.sql \cr
runSQL(sqlVar,sqlFile,pgDB)
}
