% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendSQLstatement.R
\name{sendSQLstatement}
\alias{sendSQLstatement}
\title{Send sql statement to PG}
\usage{
sendSQLstatement(sql, connList)
}
\arguments{
\item{sql}{A sql string}

\item{connList}{Named list with the following connection parameters Driver,host,user,dbname,password,port,schema}
}
\value{
no return
}
\description{
Send sql statement to PG
}
\examples{
sendSQLstatement('Create table tester as select * from test1','postgres','myDB','mypassword',5432,'prod')
}
