% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-to-df.R
\name{sql_to_df}
\alias{sql_to_df}
\title{Query PG and return results to a dataframe}
\usage{
sql_to_df(sql, pg_conn_param)
}
\arguments{
\item{sql}{A sql string}

\item{pg_conn_param}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}
}
\value{
a dataframe of the returned query
}
\description{
Query PG and return results to a dataframe
}
\examples{
sql_to_df('select * from table','localhost','postgres','myDB','mypassword',5432,'prod')
}
