% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqljoins_asstring.R
\name{sqljoins_asstring}
\alias{sqljoins_asstring}
\title{Creates a string  that can be used in a sql from clause.  Uses a standard input csv to create join}
\usage{
sqljoins_asstring(inPgTbl, schema, inCSV)
}
\arguments{
\item{inPgTbl}{Iput PG Table Name}

\item{schema}{PG database schema}

\item{inCSV}{input csv filename, note: this file must be formatted in a specific manner}
}
\value{
String
}
\description{
Creates a string  that can be used in a sql from clause.  Uses a standard input csv to create join
}
\examples{
coming soon
}
