% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif-to-gr-skey-tbl.R
\name{tif_to_gr_skey_tbl}
\alias{tif_to_gr_skey_tbl}
\title{Convert tif to grskey table}
\usage{
tif_to_gr_skey_tbl(
  src_tif_filename = "D:\\\\Projects\\\\provDataProject\\\\fadm_tfl_all_sp.tif",
  template_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  mask_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  val_field_name = "val"
)
}
\arguments{
\item{src_tif_filename}{File path to input geotiff}

\item{template_tif}{File path to template geotiff with gr_skey values}

\item{mask_tif}{File path to geotiff to be used as mask (values that ar N/a in mask will be nulled )}

\item{crop_extent}{Raster crop extent, list of c(ymin, ymax, xmin, xmax) in EPSG:3005}

\item{val_field_name}{Name of output column}
}
\value{
coming soon
}
\description{
Convert tif to grskey table
}
\examples{
coming soon
}
