% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-resultant-field-tbl.R
\name{update_resultant_field_tbl}
\alias{update_resultant_field_tbl}
\title{Update the resultant field table in PG}
\usage{
update_resultant_field_tbl(
  field_names = NULL,
  src_field_names = NULL,
  resultant_table_name = NULL,
  resultant_table_schema = NULL,
  src_attribute_table_name = NULL,
  src_attribute_table_schema = NULL,
  src_gr_skey_table = NULL,
  src_gr_skey_schema = NULL,
  out_name = resultant_table_name,
  out_schema = resultant_table_schema,
  pg_conn_param
)
}
\arguments{
\item{field_names}{vector of field names added to the resultant}

\item{src_field_names}{vector of the original field names of source table}

\item{resultant_table_name}{name of resultant table}

\item{resultant_table_schema}{name of schema of resultant table}

\item{src_gr_skey_table}{name of source gr_skey_table}

\item{src_gr_skey_schema}{name of source schema of gr_skey_table}

\item{out_name}{name of out resultant table}

\item{out_schema}{name of out resultant schema}

\item{pg_conn_param}{params}

\item{src_attribute_table}{name of source attribute table}
}
\value{
nothing is returned
}
\description{
Update the resultant field table in PG
}
\examples{
coming soon
}
