% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFKlookupPG.R
\name{updateFKlookupPG}
\alias{updateFKlookupPG}
\title{Update the foreign key lookup table in PG}
\usage{
updateFKlookupPG(
  dst_tbl,
  dst_schema,
  pk,
  suffix,
  fk_tbl,
  pg_conn_param,
  join_field = "gr_skey"
)
}
\arguments{
\item{dst_tbl}{input pg table to be incorporated into foreign key lookup table}

\item{dst_schema}{input pg table schema to be incorporated into foreign key lookup table}

\item{pk}{column to incorporated into foreign key lookup table}

\item{suffix}{suffix added to the foreign key column}

\item{fk_tbl}{current foreign key lookup table}

\item{pg_conn_param}{Named list with the following connection parameters Driver,host,user,dbname,password,port,schema}
}
\value{
nothing is returned
}
\description{
Update the foreign key lookup table in PG
}
\examples{
coming soon
}
