% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFKsrcTblpg.R
\name{updateFKsrcTblpg}
\alias{updateFKsrcTblpg}
\title{Updates Foreign Key Data Source Table in PG}
\usage{
updateFKsrcTblpg(
  data_src_tbl,
  src_type,
  src_path,
  src_lyr,
  pk,
  suffix,
  query,
  inc,
  rslt_ind,
  flds_to_keep,
  dst_tbl,
  dst_schema,
  pg_conn_param
)
}
\arguments{
\item{src_type}{data source type (one of 'gdb', 'raster','oracle','geopackage')}

\item{pk}{primary of data source e.g. objectid}

\item{suffix}{suffix used in resultant columns that were kept form the data source (e.g. "vri2021")}

\item{query}{where statement used to filter data source}

\item{inc}{1 or 0,  indicates if data source is in PG}

\item{rslt_ind}{indicates the pk is in the  pg resultant or not}

\item{flds_to_keep}{string of fields kept from data input (e.g. "feature_id,proj_age_1,live_stand_volume_125")}

\item{dst_tbl}{Data source table name}

\item{srcpath}{source path (e.g. "w:/test/tes.gdb" or 'bcgw')}

\item{srclyr}{data source  (e.g. 'veg_comp_poly' or  'whse_forest_vegetation.f_own')}

\item{nsTblm}{name of the non spatial data of data source stored in PG e.g. f_own}
}
\value{
no return
}
\description{
Updates Foreign Key Data Source Table in PG
}
\examples{
coming soon
}
