% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_batch_2_pg_grskey_grid.R
\name{add_batch_2_pg_grskey_grid}
\alias{add_batch_2_pg_grskey_grid}
\title{Update FAIB hectares database from input csv of datasets}
\usage{
add_batch_2_pg_grskey_grid(
  inCSV = "D:\\\\Projects\\\\provDataProject\\\\tools\\\\prov_data_resultant3.csv",
  connList = faibDataManagement::get_pg_conn_list(),
  oraConnList = faibDataManagement::get_ora_conn_list(),
  cropExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  gr_skey_tbl = "all_bc_gr_skey",
  wrkSchema = "whse",
  rasSchema = "raster",
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  dataSourceTblName = "data_sources",
  outTifpath = "D:\\\\Projects\\\\provDataProject",
  importrast2pg = FALSE
)
}
\arguments{
\item{inCSV}{File path to data sources csv, see inputsDatasets2load2PG.csv for example}

\item{connList}{Keyring object of Postgres credentials}

\item{oraConnList}{Keyring object of Oracle credentials}

\item{cropExtent}{list of c(ymin, ymax, xmin, xmax) in EPSG:3005}

\item{gr_skey_tbl}{Tablename of the pre-existing foreign key lookup gr_skey table. Argument to be used with suffix and rslt_ind within inCSV. Schema is assumed to be wrkSchema}

\item{wrkSchema}{Schema where gr_skey, non spatial table will be imported as well as the schema of gr_skey_tbl}

\item{rasSchema}{If importrast2pg=TRUE, schema of imported raster}

\item{grskeyTIF}{The file path to the gr_skey geotiff to be used as a template raster}

\item{maskTif}{The file path to the geotiff to be used as a mask}

\item{dataSourceTblName}{coming soon}

\item{outTifpath}{Directory where output tif if exported and where vector is temporally stored prior to import}

\item{importrast2pg}{If TRUE, raster is imported into database in rasSchema}
}
\value{
no return
}
\description{
Update FAIB hectares database from input csv of datasets
}
\examples{
coming soon
}
