% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_2_pg_grskey_grid.R
\name{add_data_2_pg_grskey_grid}
\alias{add_data_2_pg_grskey_grid}
<<<<<<< HEAD
\title{Update FAIB hectares database from input dataset}
\usage{
add_data_2_pg_grskey_grid(
  rslt_ind,
  srctype,
  srcpath,
  srclyr,
  pk,
  suffix,
  nsTblm,
  query,
  flds2keep,
=======
\title{Update FAIB hectares database from input csv of datasets}
\usage{
add_data_2_pg_grskey_grid(
  inCSV = "D:\\\\Projects\\\\provDataProject\\\\tools\\\\prov_data_resultant3.csv",
>>>>>>> 4d4bc33b907623b010db9c5c2ba86d7a04da6d64
  connList = faibDataManagement::get_pg_conn_list(),
  oraConnList = faibDataManagement::get_ora_conn_list(),
  cropExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  gr_skey_tbl = "all_bc_res_gr_skey",
  wrkSchema = "whse",
  rasSchema = "raster",
<<<<<<< HEAD
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Lands_and_Islandsincluded.tif",
=======
  templateRaster = "raster.grskey_bc_land",
>>>>>>> 4d4bc33b907623b010db9c5c2ba86d7a04da6d64
  dataSourceTblName = "data_sources",
  setwd = "D:/Projects/provDataProject",
  outTifpath = "D:\\\\Projects\\\\provDataProject",
  importrast2pg = FALSE
)
}
\arguments{
<<<<<<< HEAD
\item{rslt_ind}{coming soon}

\item{srctype}{coming soon}

\item{srcpath}{coming soon}

\item{srclyr}{coming soon}

\item{pk}{coming soon}

\item{suffix}{coming soon}

\item{nsTblm}{coming soon}

\item{query}{coming soon}

\item{flds2keep}{coming soon}
=======
\item{inCSV}{coming soon}
>>>>>>> 4d4bc33b907623b010db9c5c2ba86d7a04da6d64

\item{connList}{coming soon}

\item{cropExtent}{coming soon}

\item{gr_skey_tbl}{coming soon}

\item{wrkSchema}{coming soon}

\item{rasSchema}{coming soon}

<<<<<<< HEAD
\item{grskeyTIF}{coming soon}

\item{maskTif}{coming soon}
=======
\item{templateRaster}{coming soon}
>>>>>>> 4d4bc33b907623b010db9c5c2ba86d7a04da6d64

\item{dataSourceTblName}{coming soon}

\item{setwd}{coming soon}
}
\value{
no return
}
\description{
<<<<<<< HEAD
Update FAIB hectares database from input dataset
=======
Update FAIB hectares database from input csv of datasets
>>>>>>> 4d4bc33b907623b010db9c5c2ba86d7a04da6d64
}
\examples{
coming soon
}
