% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-add-fields-to-resultant.R
\name{batch_add_fields_to_resultant}
\alias{batch_add_fields_to_resultant}
\title{Update the resultant table with fields from previously imported datasets
A table with gr_skey needs to exist prior to running this function}
\usage{
batch_add_fields_to_resultant(
  in_csv = "batch_add_fields_to_resultant.csv",
  pg_conn_param = dadmtools::get_pg_conn_list()
)
}
\arguments{
\item{in_csv}{File path to batch add fields csv, Defaults to "batch_add_fields_to_resultant.csv"}

\item{pg_conn_param}{named list of postgres connection parameters (see get_pg_conn_list() function for more details)}
}
\value{
nothing is returned
}
\description{
Update the resultant table with fields from previously imported datasets
A table with gr_skey needs to exist prior to running this function
}
\examples{
coming soon
}
