% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-import-to-pg-gr-skey.R
\name{batch_import_to_pg_gr_skey}
\alias{batch_import_to_pg_gr_skey}
\title{Update FAIB hectares database from input csv of datasets}
\usage{
batch_import_to_pg_gr_skey(
  in_csv = "config_parameters.csv",
  out_tif_path,
  pg_conn_param = dadmtools::get_pg_conn_list(),
  ora_conn_param = dadmtools::get_ora_conn_list(),
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  gr_skey_tbl = "whse.all_bc_gr_skey",
  raster_schema = "raster",
  template_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  mask_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  data_src_tbl = "whse.data_sources",
  import_rast_to_pg = FALSE
)
}
\arguments{
\item{in_csv}{File path to data sources csv, Defaults to "config_parameters.csv"}

\item{out_tif_path}{Directory where output tif if exported and where vector is temporally stored prior to import}

\item{pg_conn_param}{Keyring object of Postgres credentials, defaults to dadmtools::get_pg_conn_list()}

\item{ora_conn_param}{Keyring object of Oracle credentials, defaults to dadmtools::get_ora_conn_list()}

\item{crop_extent}{list of c(ymin, ymax, xmin, xmax) in EPSG:3005, defaults to c(273287.5,1870587.5,367787.5,1735787.5)}

\item{gr_skey_tbl}{Schema and table name of the pre-existing gr_skey table. Argument to be used with suffix and rslt_ind within in_csv. Defaults to "whse.all_bc_gr_skey"}

\item{raster_schema}{If import_rast_to_pg = TRUE, schema of imported raster, defaults to "raster"}

\item{template_tif}{The file path to the gr_skey geotiff to be used as a template raster, fefaults to "S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\bc_01ha_gr_skey.tif"}

\item{mask_tif}{The file path to the geotiff to be used as a mask, defaults to "S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\BC_Boundary_Terrestrial.tif"}

\item{data_src_tbl}{Schema and table name of the metadata table in postgres that updates with any newly imported layer. Defaults to "whse.data_sources"}

\item{import_rast_to_pg}{If TRUE, raster is imported into database in raster_schema. Defaults to FALSE}
}
\value{
no return
}
\description{
Update FAIB hectares database from input csv of datasets
}
\examples{
coming soon
}
