% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-import-to-pg-gr-skey.R
\name{batch_import_to_pg_gr_skey}
\alias{batch_import_to_pg_gr_skey}
\title{Batch import spatial data into a gridded attribute format within PostgreSQL, structuring data according to the gr_skey grid system.}
\usage{
batch_import_to_pg_gr_skey(
  in_csv = "config_parameters.csv",
  out_tif_path,
  pg_conn_param = dadmtools::get_pg_conn_list(),
  ora_conn_param = dadmtools::get_ora_conn_list(),
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  gr_skey_tbl = "whse.all_bc_gr_skey",
  raster_schema = "raster",
  template_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  mask_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  data_src_tbl = "whse.data_sources",
  import_rast_to_pg = FALSE,
  grskey_schema = NULL
)
}
\arguments{
\item{in_csv}{File path to input configuration csv, defaults to "config_parameters.csv"}

\item{out_tif_path}{Directory where output tif is exported and where vector is temporally stored prior to import}

\item{pg_conn_param}{Keyring object of Postgres credentials, defaults to dadmtools::get_pg_conn_list()}

\item{ora_conn_param}{Keyring object of Oracle credentials, defaults to dadmtools::get_ora_conn_list()}

\item{crop_extent}{list of c(ymin, ymax, xmin, xmax) in EPSG:3005, defaults to c(273287.5,1870587.5,367787.5,1735787.5)}

\item{gr_skey_tbl}{Schema and table name of the pre-existing gr_skey table imported using function: 'import_gr_skey_tif_to_pg_rast'. Defaults to "whse.all_bc_gr_skey"}

\item{raster_schema}{If import_rast_to_pg = TRUE, schema of imported raster, defaults to "raster"}

\item{template_tif}{The file path to the gr_skey geotiff to be used as a template raster, defaults to "S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\bc_01ha_gr_skey.tif"}

\item{mask_tif}{The file path to the geotiff to be used as a mask, defaults to "S:\\FOR\\VIC\\HTS\\ANA\\workarea\\PROVINCIAL\\BC_Boundary_Terrestrial.tif"}

\item{data_src_tbl}{Schema and table name of the metadata table in postgres that updates with any newly imported layer. Defaults to "whse.data_sources"}

\item{import_rast_to_pg}{If TRUE, raster is imported into database in raster_schema. Defaults to FALSE}
}
\value{
no return
}
\description{
Function supports both vector and raster inputs. For vector data (e.g., Shapefile, FGDB, GeoPackage), the function imports the attribute table into PostgreSQL and generates a corresponding raster attribute table, where each record represents a raster pixel (one hectare). The gr_skey field acts as a globally unique primary key, while pgid links the raster attributes to the vector attributes. Each record within the raster attribute table represents one pixel.
For raster data (e.g., geotiff), the raster is required to have BC Albers coordinate reference system, (Ie. EPSG: 3005), the same grid definition as the 'template_tif' and 'mask_tif' provided to the 'import_gr_skey_tif_to_pg_rast' function and only one band. For TSR, it is recommended to use the gr_skey grid. The function imports the raster as a single attribute table. The table includes 'gr_skey', unique global cell id and the input raster pixel value. Each record represent one pixel.
}
\examples{
## After updating config_parameters.csv with source data and arguments, run the following:
batch_import_to_pg_gr_skey(
   in_csv            = 'C:\\\\<specify path>\\\\config_parameters.csv',
   out_tif_path      = 'C:\\\\<specify path>\\\\'
)
}
