% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOracleFDWpg.R
\name{createOracleFDWpg}
\alias{createOracleFDWpg}
\title{Create Foreign Data Wrapper Table in Postgres}
\usage{
createOracleFDWpg(
  dbserver,
  oraUser,
  oraPassword,
  oratable,
  connList,
  fdwSchema
)
}
\arguments{
\item{dbserver}{coming soon}

\item{oraUser}{coming soon}

\item{oraPassword}{coming soon}

\item{oratable}{coming soon}
}
\value{
None
}
\description{
Create Foreign Data Wrapper Table in Postgres
}
\examples{
coming soon
}
