% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdwTbl2PGnoSpatial.R
\name{fdwTbl2PGnoSpatial}
\alias{fdwTbl2PGnoSpatial}
\title{Create a PG non spatial table from Foreign data wrapper table}
\usage{
fdwTbl2PGnoSpatial(
  foreignTable,
  outTblName,
  pk,
  outSchema,
  connList,
  fdwSchema = "load",
  attr2keep = NULL,
  where = "",
  table_comment
)
}
\arguments{
\item{foreignTable}{coming soon}

\item{outTblName}{coming soon}

\item{pk}{coming soon}

\item{connList}{coming soon}

\item{attr2keep}{optional}

\item{schema}{coming soon}
}
\value{
None
}
\description{
Create a PG non spatial table from Foreign data wrapper table
}
\examples{
coming soon
}
