% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatePatternTableList.R
\name{getDatePatternTableList}
\alias{getDatePatternTableList}
\title{Query PG and return list of table names that are date based (named)}
\usage{
getDatePatternTableList(pattern, schema, connList, depth = 2)
}
\arguments{
\item{pattern}{An text string for the table pattern to look for date named related tables}

\item{schema}{The schema to search in}

\item{connList}{Named list with the following parameters Driver,host,user,dbname,password,port,schema}

\item{depth}{The number of matches with dates to return. default is 2}
}
\value{
a dataframe of the list of matching tables
}
\description{
Query PG and return list of table names that are date based (named)
}
\examples{
getDatePatternTableList('techpanel_scenarios','localhost','postgres','myDB','mypassword',5432,'prod')
}
