% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFDWtblSpSQL.R
\name{getFDWtblSpSQL}
\alias{getFDWtblSpSQL}
\title{Returnes a sql staement including the spatial field and primary key field of a Foreign Data Wrapper table.  Can be used in gdal rasterize statemenr}
\usage{
getFDWtblSpSQL(oratable, pk, connList, fdwSchema = "load")
}
\arguments{
\item{oratable}{coming soon}

\item{pk}{coming soon}

\item{connList}{coming soon}

\item{outTblName}{coming soon}

\item{schema}{coming soon}
}
\value{
None
}
\description{
Returnes a sql staement including the spatial field and primary key field of a Foreign Data Wrapper table.  Can be used in gdal rasterize statemenr
}
\examples{
coming soon
}
