% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFDWtblSpSQL.R
\name{getFDWtblSpSQL}
\alias{getFDWtblSpSQL}
\title{Returns a SQL statement including the spatial field and primary key field of a Foreign Data Wrapper table.  Can be used in gdal rasterize statement}
\usage{
getFDWtblSpSQL(
  dst_tbl,
  dst_schema,
  ora_tbl,
  pk,
  pg_conn_param,
  fdw_schema = "load",
  where = NULL
)
}
\arguments{
\item{dst_tbl}{coming soon}

\item{dst_schema}{coming soon}

\item{ora_tbl}{coming soon}

\item{pk}{coming soon}

\item{pg_conn_param}{coming soon}

\item{fdw_schema}{coming soon}

\item{where}{coming soon}
}
\value{
sql string
}
\description{
Returns a SQL statement including the spatial field and primary key field of a Foreign Data Wrapper table.  Can be used in gdal rasterize statement
}
\examples{
coming soon
}
