% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_skey_tif_2_pg_geom.R
\name{gr_skey_tif_2_pg_geom}
\alias{gr_skey_tif_2_pg_geom}
\title{Import Grskey TIF 2 pg table}
\usage{
gr_skey_tif_2_pg_geom(
  template_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  mask_tif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  crop_extent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  out_crop_tif_name,
  pg_conn_param = faibDataManagement::get_pg_conn_list(),
  dst_tbl = "whse.all_bc_gr_skey"
)
}
\arguments{
\item{template_tif}{The file path to the gr_skey template geotiff}

\item{mask_tif}{The file path to the geotiff to be used as a mask}

\item{crop_extent}{Raster crop extent, list of c(ymin, ymax, xmin, xmax) in EPSG:3005}

\item{out_crop_tif_name}{filename of the written output tif}

\item{pg_conn_param}{Keyring object of Postgres credentials}

\item{dst_tbl}{Postgres table name of the imported table with gr_skey and geom}
}
\value{
coming soon
}
\description{
Import Grskey TIF 2 pg table
}
\examples{
coming soon
}
