% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-blank.R
\name{is_blank}
\alias{is_blank}
\title{Find values that are blank}
\usage{
is_blank(x, false.triggers = FALSE)
}
\arguments{
\item{x}{A sql string}

\item{false.triggers}{Defaults to FALSE}
}
\value{
Boolean TRUE or FALSE
}
\description{
Find values that are blank
}
\examples{
is_blank('')
}
