% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqljoins_asstring.R
\name{sqljoins_asstring}
\alias{sqljoins_asstring}
\title{Creates FROM clause string from the standard FAIB_DATA_MANAGEMETN input csv.  Only uses inputs where inc = 1}
\usage{
sqljoins_asstring(inPgTbl, schema, inCSV)
}
\arguments{
\item{inPgTbl}{Input master gr_Key PG Table Name (containing foreign keys)}

\item{schema}{PG database schema}

\item{inCSV}{input csv filename, note: this file must be formatted in a specific manner}
}
\value{
String
}
\description{
Creates FROM clause string from the standard FAIB_DATA_MANAGEMETN input csv.  Only uses inputs where inc = 1
}
\examples{
sqljoins_asstring(masterTable,schema='postgres',inCSV='D:/inputs/inputCsv.csv')
}
