% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqljoins_asstring.R
\name{sqljoins_asstring}
\alias{sqljoins_asstring}
\title{Creates FROM clause string from the standard FAIB_DATA_MANAGEMETN input csv.  Only uses inputs where inc = 1}
\usage{
sqljoins_asstring(
  inPgTbl,
  schema,
  inDF,
  onlyIncRows = FALSE,
  filterSuffixVect = NULL
)
}
\arguments{
\item{inPgTbl}{Input master gr_Key PG Table Name (containing foreign keys)}

\item{schema}{PG database schema}

\item{inDF}{input data frame in same format as input csv}

\item{onlyIncRows}{include of rows wwhere onc field is = 1}

\item{filterSuffixVect}{vector of suffixes to be included in string}
}
\value{
String
}
\description{
Creates FROM clause string from the standard FAIB_DATA_MANAGEMETN input csv.  Only uses inputs where inc = 1
}
\examples{
sqljoins_asstring(masterTable,schema='postgres',inCSV='D:/inputs/inputCsv.csv')
}
