% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-sql-joins.R
\name{str_sql_joins}
\alias{str_sql_joins}
\title{Creates FROM clause string from the standard faib_dadm_tools input df.  Only uses inputs where inc = 1}
\usage{
str_sql_joins(
  in_pg_tbl,
  schema,
  in_df,
  only_inc_rows = FALSE,
  filter_suffix_vect = NULL
)
}
\arguments{
\item{in_pg_tbl}{Input master gr_Key PG Table Name (containing foreign keys)}

\item{schema}{PG database schema}

\item{in_df}{input data frame in same format as input csv}

\item{only_inc_rows}{include of rows wwhere onc field is = 1}

\item{filter_suffix_vect}{vector of suffixes to be included in string}
}
\value{
String
}
\description{
Creates FROM clause string from the standard faib_dadm_tools input df.  Only uses inputs where inc = 1
}
\examples{
str_sql_joins(masterTable, schema = 'postgres', inCSV = 'D:/inputs/inputCsv.csv')
}
