% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif2grskeytbl.R
\name{tif2grskeytbl}
\alias{tif2grskeytbl}
\title{Convert tif to grskey table}
\usage{
tif2grskeytbl(
  inTIF = "D:\\\\Projects\\\\provDataProject\\\\fadm_tfl_all_sp.tif",
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Lands_and_Islandsincluded.tif",
  cropExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  valueColName = "value"
)
}
\arguments{
\item{inTIF}{coming soon}

\item{grskeyTIF}{coming soon}

\item{maskTif}{coming soon}

\item{cropExtent}{coming soon}

\item{valueColName}{coming soon}
}
\value{
coming soon
}
\description{
Convert tif to grskey table
}
\examples{
coming soon
}
