% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif2grskeytbl.R
\name{tif2grskeytbl}
\alias{tif2grskeytbl}
\title{Convert tif to grskey table}
\usage{
tif2grskeytbl(
  inTIF = "D:\\\\Projects\\\\provDataProject\\\\fadm_tfl_all_sp.tif",
  grskeyTIF =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\bc_01ha_gr_skey.tif",
  maskTif =
    "S:\\\\FOR\\\\VIC\\\\HTS\\\\ANA\\\\workarea\\\\PROVINCIAL\\\\BC_Boundary_Terrestrial.tif",
  cropExtent = c(273287.5, 1870587.5, 367787.5, 1735787.5),
  valueColName = "val"
)
}
\arguments{
\item{inTIF}{File path to input geotiff}

\item{grskeyTIF}{File path to template geotiff with gr_skey values}

\item{maskTif}{File path to geotiff to be used as mask (values that ar N/a in mask will be nulled )}

\item{cropExtent}{Raster crop extent, list of c(ymin, ymax, xmin, xmax) in EPSG:3005}

\item{valueColName}{Name of output column}
}
\value{
coming soon
}
\description{
Convert tif to grskey table
}
\examples{
coming soon
}
