% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-pg-metadata-tbl.R
\name{update_pg_metadata_tbl}
\alias{update_pg_metadata_tbl}
\title{Updates Foreign Key Data Source Table in PG}
\usage{
update_pg_metadata_tbl(
  data_src_tbl,
  src_type,
  src_path,
  src_lyr,
  pk,
  query,
  inc,
  flds_to_keep,
  notes,
  overlap_ind,
  overlap_group_fields,
  dst_tbl,
  dst_schema,
  pg_conn_param
)
}
\arguments{
\item{data_src_tbl}{Data source table name}

\item{src_type}{data source type (one of 'gdb', 'raster','oracle','geopackage')}

\item{src_path}{source path (e.g. "w:/test/tes.gdb" or 'bcgw')}

\item{src_lyr}{data source  (e.g. 'veg_comp_poly' or  'whse_forest_vegetation.f_own')}

\item{pk}{primary of data source e.g. objectid}

\item{query}{where statement used to filter data source}

\item{inc}{1 or 0,  indicates if data source is in PG}

\item{flds_to_keep}{string of fields kept from data input (e.g. "feature_id,proj_age_1,live_stand_volume_125")}

\item{notes}{coming soon}

\item{overlap_ind}{True or False.  If true, it indicates that a duplicae gr_skeys will be in output gr_skey table where spatial overlaps occur.  If false, spatial overlaps will be ignored (i.e only the higher pgid value will be kept when overlaps occur)}

\item{overlap_group_fields}{The field groupings that will be used to deal with spatial overlaps. I.e. Every unique grouping of the indicated fields will be rasterized separately.}

\item{dst_tbl}{coming soon}

\item{dst_schema}{coming soon}

\item{pg_conn_param}{coming soon}
}
\value{
no return
}
\description{
Updates Foreign Key Data Source Table in PG
}
\examples{
coming soon
}
