% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFKfldTablePG.R
\name{updateFKfldTablePG}
\alias{updateFKfldTablePG}
\title{Update the foreign key field table in PG}
\usage{
updateFKfldTablePG(nsTbl, fkTBlName, suffix, connList)
}
\arguments{
\item{nsTbl}{non-spatial table to add to foreign key filed table (must have a foreign key column in the foreign key lookup table)}

\item{fkTBlName}{foreign key lookup table name}

\item{suffix}{suffix added to the foreign key column}

\item{connList}{Named list with the following connection parameters Driver,host,user,dbname,password,port,schema}
}
\value{
nothing is returned
}
\description{
Update the foreign key field table in PG
}
\examples{
coming soon
}
