% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFKlookupPG.R
\name{updateFKlookupPG}
\alias{updateFKlookupPG}
\title{Update the foreign key lookup table in PG}
\usage{
updateFKlookupPG(inLayer, pk, suffix, fkTBlName, connList, joinFld = "gr_skey")
}
\arguments{
\item{inLayer}{input pg table to be incorporated into foreign key lookup table}

\item{pk}{column to incorporated into foreign key lookup table}

\item{suffix}{suffix added to the foreign key column}

\item{fkTBlName}{current foreign key lookup table}

\item{connList}{Named list with the following connection parameters Driver,host,user,dbname,password,port,schema}
}
\value{
nothing is returned
}
\description{
Update the foreign key lookup table in PG
}
\examples{
coming soon
}
